Refactor EncryptNote component and Decrypt routes
- Extracts the encryption process to different component called
EncryptAction
- Adds new route to pass arguments to the EncryptAction component
- Changes hierarchy of decryption routes and makes payload an key both
required. This avoid situation when in form submission only the key was
entered. This resulted in url decrypt//<key>
- Removes QR code library to solve errors when decrypt link was
really big (in cases of no-store encryptions) and the QR library raised
exception. We cannot guarantee the size of the duress/decrypt link on
either store or no-store operations. Obviously in no-store operations
it's more probable to overflow this, but it can also happen in cases
where the key length is too big (too big passwords). So it's safer to
remove this feature